requirements.txt
speechrecognition
pyttsx3
pyaudio
transformers
opencv-python
pillow